pyproject.toml
[project]
name = "scratch"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.40.0",
    "mcp[cli]>=1.15.0",
    "pillow>=11.3.0",
    "playwright>=1.55.0",
]
